Update to master

+++ sparkle/src/component/mod.rs
@@ -6,5 +6,5 @@ pub trait Component: 'static + Send + Sync {}
 
 // FIXME: Change this to a more generic trait
 pub trait ComponentIndex: Component {
-    fn of(_: Option<Self>) -> uint;
+    fn of(_: Option<Self>) -> usize;
 }
